+++ app/src/main/AndroidManifest.xml
@@ -15,12 +15,12 @@
         android:theme="@style/Theme.Shopping"
         tools:targetApi="31">
         <activity
-            android:name=".ui.productdetail.ProductDetailActivity"
+            android:name=".feature.productdetail.ProductDetailActivity"
             android:exported="false"
             android:label="@string/title_activity_product_detail"
             android:theme="@style/Theme.Shopping" />
         <activity
-            android:name=".ui.cart.CartActivity"
+            android:name=".feature.cart.CartActivity"
             android:exported="false"
             android:label="@string/title_activity_cart"
             android:theme="@style/Theme.Shopping" />
