+++ app/src/main/java/woowacourse/shopping/feature/format/NumberFormatRule.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.feature.format
+
+fun interface NumberFormatRule {
+    fun apply(value: Int): String
+}
