+++ app/src/main/java/woowacourse/shopping/feature/format/PriceFormatter.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.feature.format
+
+class PriceFormatter(private val rule: NumberFormatRule, private val suffix: String) {
+    fun format(price: Int): String = "${rule.apply(price)}$suffix"
+}
