--- app/src/main/java/woowacourse/shopping/ui/format/NumberFormatRule.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.ui.format
-
-fun interface NumberFormatRule {
-    fun apply(value: Int): String
-}
