--- app/src/main/java/woowacourse/shopping/ui/format/PriceFormatter.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.ui.format
-
-class PriceFormatter(private val rule: NumberFormatRule, private val suffix: String) {
-    fun format(price: Int): String = "${rule.apply(price)}$suffix"
-}
