--- app/src/main/java/woowacourse/shopping/ui/productlist/PreviewableAsyncImage.kt
@@ -1,32 +1,0 @@
-package woowacourse.shopping.ui.productlist
-
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import coil3.compose.AsyncImage
-import coil3.request.ImageRequest
-import coil3.request.crossfade
-import woowacourse.shopping.R
-
-@Composable
-fun PreviewableAsyncImage(
-    imageUrl: String,
-    description: String,
-    modifier: Modifier = Modifier,
-) {
-    AsyncImage(
-        model = ImageRequest.Builder(LocalContext.current)
-            .data(imageUrl)
-            .crossfade(true)
-            .build(),
-        contentDescription = description,
-        contentScale = ContentScale.Crop,
-        modifier = modifier
-            .fillMaxWidth(),
-        error =
-        painterResource(R.drawable.ic_launcher_background),
-    )
-}
