--- app/src/main/java/woowacourse/shopping/ui/state/ProductUiModel.kt
@@ -1,32 +1,0 @@
-package woowacourse.shopping.ui.state
-
-import android.os.Parcelable
-import java.text.DecimalFormat
-import kotlinx.parcelize.Parcelize
-import woowacourse.shopping.ui.format.NumberFormatRule
-import woowacourse.shopping.ui.format.PriceFormatter
-
-@Parcelize
-data class ProductUiModel(val name: String, val price: String, val imageUrl: String, val id: String) : Parcelable {
-    companion object {
-        fun of(
-            name: String,
-            price: Int,
-            imageUrl: String,
-            id: String,
-        ): ProductUiModel {
-
-            val priceFormatter = PriceFormatter(
-                rule = NumberFormatRule { DecimalFormat("#,###").format(it) },
-                suffix = "원",
-            )
-
-            return ProductUiModel(
-                name = name,
-                price = priceFormatter.format(price),
-                imageUrl = imageUrl,
-                id = id,
-            )
-        }
-    }
-}
