+++ app/src/test/java/woowacourse/shopping/domain/CartTest.kt
@@ -1,6 +1,5 @@
 package woowacourse.shopping.domain
 
-import coil3.util.CoilUtils.result
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
